+++ src/main/java/com/loyalt/analytics/AnalyticsService.java
@@ -21,7 +21,6 @@ public class AnalyticsService {
     }
 
     public Analytics create(Analytics record) {
-        // Если ID транзакции не пришел, генерируем его вручную (так как ClickHouse не умеет AUTO_INCREMENT)
         if (record.getTransactionId() == null) {
             record.setTransactionId(ThreadLocalRandom.current().nextLong(1, Long.MAX_VALUE));
         }
@@ -44,15 +43,15 @@ public class AnalyticsService {
     
     String sql = 
         "SELECT " +
-        "    count() as total_trans, " + // Всего транзакций за период
-        "    count(DISTINCT client_id) as total_users, " + // Уникальных пользователей за период
+        "    count() as total_trans, " +
+        "    count(DISTINCT client_id) as total_users, " +
         "    ( " +
         "        SELECT count(DISTINCT client_id) " +
         "        FROM transactions " +
         "        WHERE partner_id = ? " +
         "        GROUP BY partner_id " +
         "        HAVING min(timestamp) >= ? AND min(timestamp) <= ? " +
-        "    ) as new_users " + // Пользователи, которые впервые появились в этот период
+        "    ) as new_users " +
         "FROM transactions " +
         "WHERE partner_id = ? AND timestamp >= ? AND timestamp <= ?";
 
@@ -60,10 +59,9 @@ public class AnalyticsService {
         return clickhouseJdbcTemplate.queryForObject(sql, (rs, rowNum) -> {
             long totalTransactions = rs.getLong("total_trans");
             long totalUsers = rs.getLong("total_users");
-            long newUsers = rs.getLong("new_users"); // Если подзапрос вернет null, JdbcTemplate вернет 0
+            long newUsers = rs.getLong("new_users");
 
-            // Контроллер ожидает объект AnalyticsDTO. 
-            // Поле date в DTO обычно заполняют текущим временем формирования отчета или датой 'to'
+
             return new AnalyticsDTO(
                     partnerId,
                     LocalDateTime.now(), 
@@ -72,12 +70,11 @@ public class AnalyticsService {
                     newUsers
             );
         }, 
-        partnerId, Timestamp.valueOf(from), Timestamp.valueOf(to), // Параметры для подзапроса new_users
-        partnerId, Timestamp.valueOf(from), Timestamp.valueOf(to)  // Параметры для основного запроса
+        partnerId, Timestamp.valueOf(from), Timestamp.valueOf(to),
+        partnerId, Timestamp.valueOf(from), Timestamp.valueOf(to)
         );
         
     } catch (Exception e) {
-        // На случай, если за период вообще не было данных, возвращаем пустой DTO, а не падаем по ошибке
         return new AnalyticsDTO(partnerId, LocalDateTime.now(), 0, 0, 0);
     }
 }
@@ -87,7 +84,7 @@ public class AnalyticsService {
         return clickhouseJdbcTemplate.query(sql, (rs, rowNum) -> new Analytics(
                 rs.getLong("transaction_id"),
                 rs.getString("partner_id"),
-                rs.getLong("client_id"),
+                rs.getInt("client_id"),
                 rs.getLong("amount"),
                 rs.getTimestamp("timestamp").toLocalDateTime()
         ), limit);
@@ -99,7 +96,7 @@ public class AnalyticsService {
             Analytics record = clickhouseJdbcTemplate.queryForObject(sql, (rs, rowNum) -> new Analytics(
                     rs.getLong("transaction_id"),
                     rs.getString("partner_id"),
-                    rs.getLong("client_id"),
+                    rs.getInt("client_id"),
                     rs.getLong("amount"),
                     rs.getTimestamp("timestamp").toLocalDateTime()
             ), id);
@@ -110,7 +107,6 @@ public class AnalyticsService {
     }
 
     public void deleteTransaction(Long id) {
-        // В ClickHouse удаление идеологически асинхронно через мутации (ALTER TABLE ... DELETE)
         String sql = "ALTER TABLE transactions DELETE WHERE transaction_id = ?";
         clickhouseJdbcTemplate.execute(sql);
     }
